+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes of the PHP Mode 1.19.1 release series are documented in this
 
  * Make obsolete `php-mode-version-number` contstant variable ([#712])
    * `(php-mode-version :as-number t)` is provided for use cases comparing as versions, but generally SHOULD NOT be dependent on the PHP Mode version.
+ * Make obsolete `php-mode-disable-c-mode-hook` customize variable ([#718])
 
 ### Fixed
 
@@ -44,6 +45,7 @@ All notable changes of the PHP Mode 1.19.1 release series are documented in this
 [#715]: https://github.com/emacs-php/php-mode/pull/715
 [#716]: https://github.com/emacs-php/php-mode/pull/716
 [#717]: https://github.com/emacs-php/php-mode/pull/717
+[#718]: https://github.com/emacs-php/php-mode/pull/718
 
 ## [1.24.1] - 2022-10-08